requirements.txt
# Bot dependencies
aiogram>=3.0.0
apscheduler>=3.10.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
httpx>=0.25.0
aiosqlite>=0.19.0

# API dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Additional dependencies for integration
aiohttp>=3.8.0
